Move Python code to src/native/
